+++ common/src/main/java/com/alien/client/animation/entity/OvomorphAnimator.java
@@ -13,8 +13,7 @@ import com.avp.AVPResources;
 
 public class OvomorphAnimator extends AzEntityAnimator<Ovomorph> {
 
-    // TODO: Change this to "ovomorph" with 0.2.0.
-    private static final String NAME = "ovamorph";
+    private static final String NAME = "ovomorph";
 
     private static final ResourceLocation ANIMATION = AVPResources.entityAnimationLocation(NAME);