+++ common/src/main/java/com/avp/common/registry/init/AVPSoundEvents.java
@@ -13,17 +13,13 @@ public class AVPSoundEvents {
 
     public static final AVPDeferredHolder<SoundEvent> BLOCK_RESIN_SPREAD = register("block.resin.spread");
 
-    // TODO: Change this to "ovomorph" with 0.2.0.
-    public static final AVPDeferredHolder<SoundEvent> ENTITY_OVOMORPH_HATCH = register("entity.ovamorph.hatch");
+    public static final AVPDeferredHolder<SoundEvent> ENTITY_OVOMORPH_HATCH = register("entity.ovomorph.hatch");
 
-    // TODO: Change this to "ovomorph" with 0.2.0.
-    public static final AVPDeferredHolder<SoundEvent> ENTITY_OVOMORPH_LAID = register("entity.ovamorph.laid");
+    public static final AVPDeferredHolder<SoundEvent> ENTITY_OVOMORPH_LAID = register("entity.ovomorph.laid");
 
-    // TODO: Change this to "ovomorph" with 0.2.0.
-    public static final AVPDeferredHolder<SoundEvent> ENTITY_OVOMORPH_ROOT = register("entity.ovamorph.root");
+    public static final AVPDeferredHolder<SoundEvent> ENTITY_OVOMORPH_ROOT = register("entity.ovomorph.root");
 
-    // TODO: Change this to "ovomorph" with 0.2.0.
-    public static final AVPDeferredHolder<SoundEvent> ENTITY_OVOMORPH_SHEAR = register("entity.ovamorph.shear");
+    public static final AVPDeferredHolder<SoundEvent> ENTITY_OVOMORPH_SHEAR = register("entity.ovomorph.shear");
 
     public static final AVPDeferredHolder<SoundEvent> ENTITY_QUEEN_ARM_ATTACK = register("entity.queen.arm_attack");